requirements.txt
Django
psycopg2-binary
gunicorn
whitenoise
Pillow
